lab_2/requirements.txt
fastapi
uvicorn[standard]
kafka-python
psycopg2-binary
pydantic
python-dotenv
